backend/requirements.txt
anyio==4.9.0
babel==2.17.0
bcrypt==3.2.0
blinker==1.8.2
cachetools==6.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
cryptography==3.4.8
dnspython==2.7.0
exceptiongroup==1.3.0
fasteners==0.14.1
Flask==3.0.3
flask-cors==6.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
numpy==1.22.4
onnxruntime==1.19.0
opencv-python==4.11.0.86
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
Pygments==2.19.1
pymongo==4.13.0
python-dotenv==1.1.0
pytesseract==0.3.13
requests==2.25.1
rich==14.0.0
scipy==1.13.0
seaborn==0.13.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==1.26.5
Werkzeug==3.0.3
plastex
flask
